fix(ci): resolve format and binary E2E test failures
- Run deno fmt on workflow/index.ts, html/utils.ts, agent-card.tsx
- Inline Link component in router barrel (avoids relative import in
  embedded framework sources that breaks compiled binary SSR resolution)
- Delete components/Link.tsx (zero consumers), remove from components barrel
- Link now lives exclusively in veryfront/router
