+++ cli/templates/manifest.json
