--- src/react/components/Link.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-export type LinkProps = React.AnchorHTMLAttributes<HTMLAnchorElement> & {
-  prefetch?: boolean;
-};
-
-export function Link({
-  prefetch = true,
-  children,
-  ...rest
-}: LinkProps): React.ReactElement {
-  return (
-    <a {...rest} data-prefetch={prefetch ? "true" : undefined}>
-      {children}
-    </a>
-  );
-}
