+++ src/react/components/index.ts
@@ -1,5 +1,4 @@
 export { AppWrapper, type AppWrapperProps } from "./AppWrapper.tsx";
 export { Head } from "./Head.tsx";
-export { Link, type LinkProps } from "./Link.tsx";
 export { MDXProvider, type MDXProviderProps, useMDXComponents } from "./MDXProvider.tsx";
 export * from "./optimized-image/index.ts";
